requirements.txt
# Core dependencies for WAF ingestion and query
llama-index>=0.10.0
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0
html2text>=2020.1.16
requests>=2.31.0
openai>=1.12.0
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0

# Optional: For better HTML parsing
lxml>=4.9.0
